Apply Rename Refactoring
