+++ Descension/Assets/Prefab/ArrowPrefab.prefab
@@ -11,8 +11,8 @@ GameObject:
   - component: {fileID: 7911246415520734235}
   - component: {fileID: 7911246415520734234}
   - component: {fileID: 7911246415520734229}
-  - component: {fileID: 7911246415520734232}
   - component: {fileID: 7911246415520734228}
+  - component: {fileID: 1354975813824798516}
   m_Layer: 0
   m_Name: ArrowPrefab
   m_TagString: Untagged
@@ -106,20 +106,6 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 0
---- !u!114 &7911246415520734232
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7911246415520734231}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 95caccdcca40e264db1fccb5fef8ba6d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  speed: 8
-  body: {fileID: 7911246415520734229}
 --- !u!58 &7911246415520734228
 CircleCollider2D:
   m_ObjectHideFlags: 0
@@ -136,3 +122,18 @@ CircleCollider2D:
   m_Offset: {x: 0, y: 0}
   serializedVersion: 2
   m_Radius: 0.5
+--- !u!114 &1354975813824798516
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7911246415520734231}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 95caccdcca40e264db1fccb5fef8ba6d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  speed: 8
+  damage: 10
+  body: {fileID: 0}
