+++ Descension/Assets/Scripts/Actor/AI/enemyOne.cs
@@ -1,73 +1,75 @@
-using System.Collections;
-using System.Collections.Generic;
+using Actor.Player;
 using UnityEngine;
 
-public class enemyOne : MonoBehaviour
+namespace Actor.AI
 {
-    public float hitPoints = 100f;
-    public float damage = 15f;
-
-    public GameObject floatingTextDialogue;
-    public GameObject floatingTextDamage;
-
-    Vector3 movement;
-
-    float pointA = 66f;
-    float pointB = 67f;
-
-    float patrolTarget;
-
-    // Start is called before the first frame update
-    void Start()
+    public class EnemyOne : MonoBehaviour
     {
-        patrolTarget = pointB;
-    }
+        public float hitPoints = 100f;
+        public float damage = 15f;
 
-    // Update is called once per frame
-    void Update()
-    {
-        if (this.hitPoints < 0) {
-            // this.showFloatingTextDialogue("DEAD!");
-            Destroy(gameObject);
-        }
+        public GameObject floatingTextDialogue;
+        public GameObject floatingTextDamage;
 
-        // Debug.Log(transform.position.x);
+        Vector3 movement;
 
-        if (patrolTarget == pointB) {
-            movement = new Vector3(2f, 0f, 0f);
-        } else if (patrolTarget == pointA) {
-            movement = new Vector3(-2f, 0f, 0f);
-        }
+        float pointA = 66f;
+        float pointB = 67f;
 
-        transform.position += movement * Time.deltaTime;
+        float patrolTarget;
 
-        if (patrolTarget == pointB && transform.position.x >= pointB) {
-            patrolTarget = pointA;
-        } 
-        if (patrolTarget == pointA && transform.position.x <= pointA) {
+        // Start is called before the first frame update
+        void Start()
+        {
             patrolTarget = pointB;
         }
-    }
 
-    void OnCollisionEnter2D(Collision2D collision) {
-        if (collision.gameObject.name == "Player") {
-            FindObjectOfType<player>().inflictDamage(this.damage);
-            // this.showFloatingTextDialogue("AAAW");
+        // Update is called once per frame
+        void Update()
+        {
+            if (this.hitPoints < 0) {
+                // this.showFloatingTextDialogue("DEAD!");
+                Destroy(gameObject);
+            }
+
+            // Debug.Log(transform.position.x);
+
+            if (patrolTarget == pointB) {
+                movement = new Vector3(2f, 0f, 0f);
+            } else if (patrolTarget == pointA) {
+                movement = new Vector3(-2f, 0f, 0f);
+            }
+
+            transform.position += movement * Time.deltaTime;
+
+            if (patrolTarget == pointB && transform.position.x >= pointB) {
+                patrolTarget = pointA;
+            } 
+            if (patrolTarget == pointA && transform.position.x <= pointA) {
+                patrolTarget = pointB;
+            }
         }
-    }
 
-    public void inflictDamage(float damage) {
-        this.hitPoints -= damage;
-        showFloatingTextDamage("HP -" + damage.ToString());
-    }
+        void OnCollisionEnter2D(Collision2D collision) {
+            if (collision.gameObject.name == "Player") {
+                FindObjectOfType<PlayerController>().inflictDamage(this.damage);
+                // this.showFloatingTextDialogue("AAAW");
+            }
+        }
 
-    void showFloatingTextDamage(string text) {
-        var t = Instantiate(floatingTextDamage, transform.position, Quaternion.identity);
-        t.GetComponent<TextMesh>().text = text;
-    }
+        public void inflictDamage(float damage) {
+            this.hitPoints -= damage;
+            showFloatingTextDamage("HP -" + damage.ToString());
+        }
 
-    void showFloatingTextDialogue(string text) {
-        var t = Instantiate(floatingTextDialogue, transform.position, Quaternion.identity);
-        t.GetComponent<TextMesh>().text = text;
+        void showFloatingTextDamage(string text) {
+            var t = Instantiate(floatingTextDamage, transform.position, Quaternion.identity);
+            t.GetComponent<TextMesh>().text = text;
+        }
+
+        void showFloatingTextDialogue(string text) {
+            var t = Instantiate(floatingTextDialogue, transform.position, Quaternion.identity);
+            t.GetComponent<TextMesh>().text = text;
+        }
     }
 }
