+++ Descension/Assets/Scripts/Actor/Player/PlayerController2.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: aa2369727f2a30a49b204240d6ab4077
+guid: 679c88b4157a801449ad8107adc65a58
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
