--- Descension/Assets/Scripts/Actor/Player/player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 72855056329b41e4780ee5330cd7d3fc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
