+++ Descension/Assets/Scripts/Environment/floatingText.cs
@@ -1,18 +1,19 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
-public class floatingText : MonoBehaviour
+namespace Environment
 {
-    public float duration = 1f;
+    public class FloatingText : MonoBehaviour
+    {
+        public float duration = 1f;
 
-    // Start is called before the first frame update
-    void Start() {
-        Destroy(gameObject, duration);
-    }
+        // Start is called before the first frame update
+        void Start() {
+            Destroy(gameObject, duration);
+        }
 
-    // Update is called once per frame
-    void Update() {
+        // Update is called once per frame
+        void Update() {
         
+        }
     }
 }
