+++ Descension/Assets/Scripts/Environment/levelTwoHole.cs
@@ -1,39 +1,30 @@
-using System.Collections;
-using System.Collections.Generic;
+using Actor.Player;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class levelTwoHole : MonoBehaviour
+namespace Environment
 {
-    public Image dialogueBox;
-    public Text dialogueText;
-
-    // Start is called before the first frame update
-    void Start() {
-        
-    }
-
-    // Update is called once per frame
-    void Update() {
-        
-    }
-
-    void OnCollisionEnter2D(Collision2D collision) {
-        // Debug.Log("collided");
-        if (collision.gameObject.name == "Player") {
-            if (FindObjectOfType<player>().ropeQuantity > 0) {
-                FindObjectOfType<player>().addRope(-1);
-                showText("Descend to level two...");
-            } else {
-                showText("You need a rope in order to descend");
+    public class LevelTwoHole : MonoBehaviour
+    {
+        public Image dialogueBox;
+        public Text dialogueText;
+    
+        void OnCollisionEnter2D(Collision2D collision) {
+            // Debug.Log("collided");
+            if (collision.gameObject.name == "Player") {
+                if (FindObjectOfType<PlayerController>().ropeQuantity > 0) {
+                    FindObjectOfType<PlayerController>().addRope(-1);
+                    showText("Descend to level two...");
+                } else {
+                    showText("You need a rope in order to descend");
+                }
             }
-
         }
-    }
 
-    void showText(string text) {
-        dialogueBox.enabled = true;
-        dialogueText.enabled = true;
-        dialogueText.text = text;
+        void showText(string text) {
+            dialogueBox.enabled = true;
+            dialogueText.enabled = true;
+            dialogueText.text = text;
+        }
     }
 }
