+++ Descension/Assets/Scripts/Items/Arrow.cs
@@ -1,30 +1,35 @@
-using System.Collections;
-using System.Collections.Generic;
 using Actor.AI;
 using UnityEngine;
 
-public class arrow : MonoBehaviour
+namespace Items
 {
-    public float speed = 8;
-    public float damage = 10;
-    public Rigidbody2D body; 
-    
-    // Start is called before the first frame update
-    void Start() {
+    [RequireComponent(typeof(Rigidbody2D))]
+    public class Arrow : MonoBehaviour
+    {
+        // TODO: Given that this is an entity and not a consumable item,
+        // TODO: this class should be moved elsewhere later.
+
+        public float speed = 8;
+        public float damage = 10;
+        public Rigidbody2D body; 
         
-    }
+        void Awake()
+        {
+            body = GetComponent<Rigidbody2D>();
+        }
 
-    // Update is called once per frame
-    void Update() {
-        body.velocity = transform.right * speed;
-    }
+        // Update is called once per frame
+        void Update() {
+            body.velocity = transform.right * speed;
+        }
 
-    void OnTriggerEnter2D(Collider2D collision) {
-        // set "Enemy" Tag to enemy object for this to work
-        if (collision.CompareTag("Enemy")) {
-            // Debug.Log("attacked enemy");
-            collision.gameObject.GetComponent<AIController>().InflictDamage(this.damage);
-            Destroy(gameObject);
+        void OnTriggerEnter2D(Collider2D collision) {
+            // set "Enemy" Tag to enemy object for this to work
+            if (collision.CompareTag("Enemy")) {
+                // Debug.Log("attacked enemy");
+                collision.gameObject.GetComponent<AIController>().InflictDamage(this.damage);
+                Destroy(gameObject);
+            }
         }
     }
 }
