+++ Descension/Assets/Scripts/Items/bowItem.cs
@@ -1,34 +1,26 @@
-using System.Collections;
-using System.Collections.Generic;
+using Actor.Player;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class bowItem : MonoBehaviour
+namespace Items
 {
-    public Image dialogueBox;
-    public Text dialogueText;
-
-    // Start is called before the first frame update
-    void Start() {
+    public class BowItem : MonoBehaviour
+    {
+        public Image dialogueBox;
+        public Text dialogueText;
         
-    }
-
-    // Update is called once per frame
-    void Update() {
-        
-    }
-
-    void OnCollisionEnter2D(Collision2D collision) {
-        if (collision.gameObject.name == "Player") {
-            FindObjectOfType<player>().addBow();
-            showText("Bow collected");
-            Destroy(gameObject);
+        void OnCollisionEnter2D(Collision2D collision) {
+            if (collision.gameObject.name == "Player") {
+                FindObjectOfType<PlayerController>().addBow();
+                showText("Bow collected");
+                Destroy(gameObject);
+            }
         }
-    }
 
-    void showText(string text) {
-        dialogueBox.enabled = true;
-        dialogueText.enabled = true;
-        dialogueText.text = text;
+        void showText(string text) {
+            dialogueBox.enabled = true;
+            dialogueText.enabled = true;
+            dialogueText.text = text;
+        }
     }
 }
