+++ Descension/Assets/Scripts/Items/bowItem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 823997f47fe34e2489b698b3fcace61e
+guid: 27bc5f9e4f21f7e4197c3be8b9cd60bd
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
