+++ Descension/Assets/Scripts/Items/ropeItem.cs
@@ -1,35 +1,27 @@
-using System.Collections;
-using System.Collections.Generic;
+using Actor.Player;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class ropeItem : MonoBehaviour
+namespace Items
 {
-    public Image dialogueBox;
-    public Text dialogueText;
-    public float quantity = 1;
-
-    // Start is called before the first frame update
-    void Start() {
+    public class RopeItem : MonoBehaviour
+    {
+        public Image dialogueBox;
+        public Text dialogueText;
+        public float quantity = 1;
         
-    }
-
-    // Update is called once per frame
-    void Update() {
-        
-    }
-
-    void OnCollisionEnter2D(Collision2D collision) {
-        if (collision.gameObject.name == "Player") {
-            FindObjectOfType<player>().addRope(this.quantity);
-            showText("Rope Collected");
-            Destroy(gameObject);
+        void OnCollisionEnter2D(Collision2D collision) {
+            if (collision.gameObject.name == "Player") {
+                FindObjectOfType<PlayerController>().addRope(this.quantity);
+                showText("Rope Collected");
+                Destroy(gameObject);
+            }
         }
-    }
 
-    void showText(string text) {
-        dialogueBox.enabled = true;
-        dialogueText.enabled = true;
-        dialogueText.text = text;
+        void showText(string text) {
+            dialogueBox.enabled = true;
+            dialogueText.enabled = true;
+            dialogueText.text = text;
+        }
     }
 }
