+++ Descension/Assets/Scripts/Items/torchItem.cs
@@ -1,35 +1,27 @@
-using System.Collections;
-using System.Collections.Generic;
+using Actor.Player;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class torchItem : MonoBehaviour
+namespace Items
 {
-    public Image dialogueBox;
-    public Text dialogueText;
-    public float quantity = 20;
-
-    // Start is called before the first frame update
-    void Start() {
+    public class TorchItem : MonoBehaviour
+    {
+        public Image dialogueBox;
+        public Text dialogueText;
+        public float quantity = 20;
         
-    }
-
-    // Update is called once per frame
-    void Update() {
-        
-    }
-
-    void OnCollisionEnter2D(Collision2D collision) {
-        if (collision.gameObject.name == "Player") {
-            FindObjectOfType<player>().addTorch(this.quantity);
-            showText("Torch Collected");
-            Destroy(gameObject);
+        void OnCollisionEnter2D(Collision2D collision) {
+            if (collision.gameObject.name == "Player") {
+                FindObjectOfType<PlayerController>().addTorch(this.quantity);
+                showText("Torch Collected");
+                Destroy(gameObject);
+            }
         }
-    }
 
-    void showText(string text) {
-        dialogueBox.enabled = true;
-        dialogueText.enabled = true;
-        dialogueText.text = text;
+        void showText(string text) {
+            dialogueBox.enabled = true;
+            dialogueText.enabled = true;
+            dialogueText.text = text;
+        }
     }
 }
